fix: update test location
