+++ tests/mocks/index.ts
@@ -77,3 +77,45 @@ export const TEST_WALLET: Wallet = {
     },
   ],
 };
+
+export const TEST_ACCOUNTS_WITH_ADDRESS = [
+  {
+    stxPrivateKey: 'cc440106121617d5ea23646552aa40a62316878293a1baa7a02fbb071b8cbe2901',
+    dataPrivateKey: '504b91ea465bb9f6c28a38432869bd04dbeb85b46525bcf5e9615a9ddfcc178c',
+    appsKey:
+      'xprvA2KzZmgR2PMM55CgbWxRC6K74n4ANrBitjxKnKmEJYLq4tfQX5p17NpUx9gvk4fnnJDym14E1hixmDzjfbt4Fmf73g1RcvQZpMzih81Lg1S',
+    salt: 'e95b8fed404c4130267b15be0df5aecb7c118f1297bf25d096e2bf442ec7b1a7',
+    index: 0,
+    address: 'SP5ZJW9Z6JJ2N86A168AR1M7CVCGZRWZSW6SEM6H',
+  },
+  {
+    stxPrivateKey: '1499c9194feed4cbca3bd9f7a2b3413684d30c1a9d8a7f044398f9e27ae5d00301',
+    dataPrivateKey: 'f96a1e65ce95db1ee6ff2e53988d4a3524c433548758d95354ea05fc24cc4505',
+    appsKey:
+      'xprvA1eQc5DGP8KU275yPdXUbWXPW4Ba2h2poSpXnV5jxxfpPghuaP4p8R1YHg8MsmqGpwFQU4R74cU5x8E7dExQN14cvPHn6c1vxajxWjQxTpM',
+    salt: 'e95b8fed404c4130267b15be0df5aecb7c118f1297bf25d096e2bf442ec7b1a7',
+    index: 1,
+    username: 'fdsfdsfdf.id.blockstack',
+    address: 'SP2PHCPANVT8DVPSY5W2ZZ81M285Q5Z8Y6FQ2ZH9D',
+  },
+  {
+    stxPrivateKey: 'b8e5b2c33be621fb16a5260b30d526cadc7087d58736f2f3c6f84d31be452b6601',
+    dataPrivateKey: 'bdcc48a4b8c098a7cb06e7221ee0b784c05083353dad964674a4eebfaa104f0d',
+    appsKey:
+      'xprvA1vvF4rkEyBDLuvQKTa9YvCNKHcTNsrBwD5imsnnatG9RiyF5634o2GBwb5YvjC237LK6k273xRmaiCsNKpbS2EXUGPTtWo3CBFKRmzEHti',
+    salt: 'e95b8fed404c4130267b15be0df5aecb7c118f1297bf25d096e2bf442ec7b1a7',
+    index: 2,
+    username: 'thisis45678.id.blockstack',
+    address: 'SP35Z3YQCTC1WZ8Z7AKHGE91HK05WKMKPTMATGQYP',
+  },
+  {
+    stxPrivateKey: 'da62811e06fe6fb394982a740fdcab6f1194ed85c5f6421021777638512d41ee01',
+    dataPrivateKey: '522a6fb27d2255c30ab7d83ffc3b43cda4627c505619f0bb1e7de66c297bc53c',
+    appsKey:
+      'xprvA23n53oLHrMsQcVULk6UBg3PJUWHC3L6W8DmKqzGEVprqjSF1qs8hcZfP3H1zg1asSUDDZBXNw81p5BBALZorQz2fKNeWGfwmRhUsbMiVsH',
+    salt: 'e95b8fed404c4130267b15be0df5aecb7c118f1297bf25d096e2bf442ec7b1a7',
+    index: 3,
+    username: 'integration_tester.test-personal.id',
+    address: 'SP23C2PBK6PVQEW2M17JSB6S2K0VZ611PPJ0D20NH',
+  },
+];
